Double changed to Long for Double is rounding large numbers

+++ src/main/java/uz/backall/profit/profitGroup/ProfitGroupCreateDTO.java
@@ -16,5 +16,5 @@ public class ProfitGroupCreateDTO {
   private Long storeId;
 
   @NotNull
-  private Double profit;
+  private Long profit;
 }
